+++ docs/changelog.rst
@@ -21,7 +21,7 @@ Major change
 Minor change
 ~~~~~~~~~~~~
 
-* Add :func:`~capytaine.io.xarray.assemble_matrices` function which is a simplified version of `~capytaine.io.xarray.assemble_dataset` without metadata, meant to be used mostly for teaching.
+* Add :func:`~capytaine.io.xarray.assemble_matrices` function which is a simplified version of `~capytaine.io.xarray.assemble_dataset` without metadata, meant to be used mostly for teaching. (:pull:`643`)
 
 Bug fixes
 ~~~~~~~~~