+++ docs/changelog.rst
@@ -30,6 +30,7 @@ Bug fixes
 
 * Fix missing geometric center in legacy predefined body :class:`~capytaine.bodies.predefined.rectangles.ReflectionSymmetricMesh`. It was causing inconsistent definition of dofs with respect to earlier versions. (:pull:`625`)
 
+* Fix Python implementation of the Prony decomposition for the finite depth Green function. The default is still the legacy Fortran implementation. (:pull:`621`).
 
 ---------------------------------
 New in version 2.2.1 (2024-11-18)